Added Airing shows
Scripts I wrote for Judas, includes weekly scripts End of Season batches, and a couple odd Web source encodes.
Digimon doesn't have any BDs so we used web, and I've kept it here till we v2 it with BDs. The One Piece script is being used to encode currently airing epsiodes, and it's probably never going to stop airing so it's here.